chore: updates from code-review
